Fix not being able to connect from the server list whilst already connected.
This pull request fixes an issue where if a user is already connected to a server, they must disconnect first via the disconnect button before connecting to another. It now will allow the user to click the connect button and it will disconnect for them first and then continue connecting to the selected server.